Move our DWARF parsing back to a single thread, which is required since Go uses DW_AT_abstract_origin to keep track of DIEs global offsets throughout the file.
This is particularly sad since Go would stand to benefit a lot from
a multi-threaded DWARF parser, but I can't see a way to get this done
other than this. We may be able to parallelize some of the more
trivial parts of the parser, but at that point it's not even worth it.